requirements.txt
streamlit
rembg
pillow
numpy
opencv-python
onnxruntime
streamlit-image-comparison
